requirements.txt
requests
pdf2image
pyzbar
pillow
pytesseract
opencv-python
flask